Add an enhancement for a minimal ui.
This removes the button backgrounds, hides the hearts when they are full and the magic when its not in use. It also hides the rupee / key counters. All this is still visible on the pause screen or when the minimap is visible. It also makes the minimap off by default, but you can still turn it on.